+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- Fixed formatting in `gantry cluster` and `gantry config` commands.
+
 ## [v1.0.0](https://github.com/allenai/beaker-gantry/releases/tag/v1.0.0) - 2024-05-24
 
 ### Added